feat: shopId, 날짜로 매출 조회 기능 구현

+++ src/main/java/uos/uos25/employee/service/EmployeeService.java
@@ -156,7 +156,7 @@ public class EmployeeService {
                         .mapToLong(
                                 employeeWorkingHistory -> employeeWorkingHistory.getWorkingHour())
                         .sum();
-        Long salary = employee.calculateSalary(totalWorkingHour);
+        Integer salary = employee.calculateSalary(totalWorkingHour);
 
         Disbursement disbursement =
                 disbursementService.create(employee, salary, salaryCalculationRequestDTO.getDate());
